+++ packages/bloc_test/lib/src/when_listen.dart
@@ -7,6 +7,9 @@ import 'package:mockito/mockito.dart';
 /// Use [whenListen] if you want to return a canned `Stream` of states
 /// for a [bloc] instance.
 ///
+/// [whenListen] also handles stubbing the `state` of the bloc to stay
+/// in sync with the emitted state.
+///
 /// Return a canned state stream of `[0, 1, 2, 3]`
 /// when `counterBloc.listen` is called.
 ///
@@ -18,21 +21,28 @@ import 'package:mockito/mockito.dart';
 /// is the canned `Stream`.
 ///
 /// ```dart
-/// expectLater(
+/// await expectLater(
 ///   counterBloc,
 ///   emitsInOrder(
 ///     <Matcher>[equals(0), equals(1), equals(2), equals(3), emitsDone],
 ///   )
-/// )
+/// );
+/// expect(counterBloc.state, equals(3));
 /// ```
 void whenListen<Event, State>(
   Bloc<Event, State> bloc,
   Stream<State> stream,
 ) {
+  StreamSubscription<State> subscription;
   final broadcastStream = stream.asBroadcastStream();
+  subscription = broadcastStream.listen(
+    (state) => when(bloc.state).thenReturn(state),
+    onDone: () => subscription?.cancel(),
+  );
   when(bloc.skip(any)).thenAnswer(
-    (invocation) =>
-        broadcastStream.skip(invocation.positionalArguments.first as int),
+    (invocation) => broadcastStream.skip(
+      invocation.positionalArguments.first as int,
+    ),
   );
 
   when(bloc.listen(