+++ packages/bloc_test/test/when_listen_test.dart
@@ -29,6 +29,60 @@ void main() {
       );
     });
 
+    test('can mock the stream of a single bloc with delays', () async {
+      final counterBloc = MockCounterBloc();
+      final controller = StreamController<int>();
+      whenListen(counterBloc, controller.stream);
+      expectLater(
+        counterBloc,
+        emitsInOrder(
+          <Matcher>[equals(0), equals(1), equals(2), equals(3), emitsDone],
+        ),
+      );
+      controller.add(0);
+      await Future.delayed(const Duration(milliseconds: 10));
+      controller.add(1);
+      await Future.delayed(const Duration(milliseconds: 10));
+      controller.add(2);
+      await Future.delayed(const Duration(milliseconds: 10));
+      controller.add(3);
+      controller.close();
+    });
+
+    test('can mock the state of a single bloc with delays', () async {
+      final counterBloc = MockCounterBloc();
+      final controller = StreamController<int>();
+      whenListen(counterBloc, controller.stream);
+      expectLater(
+        counterBloc,
+        emitsInOrder(
+          <Matcher>[equals(0), equals(1), equals(2), equals(3), emitsDone],
+        ),
+      ).then((_) {
+        expect(counterBloc.state, equals(3));
+      });
+      controller.add(0);
+      await Future.delayed(const Duration(milliseconds: 10));
+      controller.add(1);
+      await Future.delayed(const Duration(milliseconds: 10));
+      controller.add(2);
+      await Future.delayed(const Duration(milliseconds: 10));
+      controller.add(3);
+      controller.close();
+    });
+
+    test('can mock the state of a single bloc', () async {
+      final counterBloc = MockCounterBloc();
+      whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
+      await expectLater(
+        counterBloc,
+        emitsInOrder(
+          <Matcher>[equals(0), equals(1), equals(2), equals(3), emitsDone],
+        ),
+      );
+      expect(counterBloc.state, equals(3));
+    });
+
     test('can mock the stream of a single bloc as broadcast stream', () {
       final counterBloc = MockCounterBloc();
       whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
@@ -57,6 +111,18 @@ void main() {
       );
     });
 
+    test('can mock the state of a single bloc with skip(1)', () async {
+      final counterBloc = MockCounterBloc();
+      whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
+      await expectLater(
+        counterBloc.skip(1),
+        emitsInOrder(
+          <Matcher>[equals(1), equals(2), equals(3), emitsDone],
+        ),
+      );
+      expect(counterBloc.state, equals(3));
+    });
+
     test('can mock the stream of a single bloc with skip(2)', () {
       final counterBloc = MockCounterBloc();
       whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
@@ -66,6 +132,18 @@ void main() {
           <Matcher>[equals(2), equals(3), emitsDone],
         ),
       );
+    });
+
+    test('can mock the state of a single bloc with skip(2)', () async {
+      final counterBloc = MockCounterBloc();
+      whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
+      await expectLater(
+        counterBloc.skip(2),
+        emitsInOrder(
+          <Matcher>[equals(2), equals(3), emitsDone],
+        ),
+      );
+      expect(counterBloc.state, equals(3));
     });
 
     test('can mock the stream of a single bloc with skip(3)', () {
@@ -79,11 +157,31 @@ void main() {
       );
     });
 
+    test('can mock the state of a single bloc with skip(3)', () async {
+      final counterBloc = MockCounterBloc();
+      whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
+      await expectLater(
+        counterBloc.skip(3),
+        emitsInOrder(
+          <Matcher>[equals(3), emitsDone],
+        ),
+      );
+      expect(counterBloc.state, equals(3));
+    });
+
     test('can mock the stream of a bloc dependency', () {
       final counterBloc = MockCounterBloc();
       whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
       final sumBloc = SumBloc(counterBloc);
       expectLater(sumBloc, emitsInOrder(<int>[0, 1, 3, 6]));
+    });
+
+    test('can mock the state of a bloc dependency', () async {
+      final counterBloc = MockCounterBloc();
+      whenListen(counterBloc, Stream.fromIterable([0, 1, 2, 3]));
+      final sumBloc = SumBloc(counterBloc);
+      await expectLater(sumBloc, emitsInOrder(<int>[0, 1, 3, 6]));
+      expect(sumBloc.state, equals(6));
     });
   });
 }