+++ packages/bloc_test/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 5.1.0
 
 - Add `errors` to `blocTest` to enable expecting unhandled exceptions within blocs.
+- Update `whenListen` to also handle stubbing the state property of the bloc.
 
 # 5.0.0